requirements.txt
torch<3
numpy<3
hydra-core
tqdm
datasets
torchvision
matplotlib
diffusers
accelerate